requirements.txt
flask
flask-cors
pymongo
python-dotenv
sendgrid
